# server/docs/swagger.yaml
basePath: /api
definitions:
...
  httpx.ErrorDetail:
    properties:
      location:
        type: string
      message:
        type: string
    type: object
  httpx.ProblemDetails:
    properties:
      code:
        type: string
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/httpx.ErrorDetail'
        type: array
      instance:
        type: string
      requestId:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for the SITCON Camp 2026 game.
  title: Camp 2026 Game API
  version: 0.1.0
paths:
...
  /me/events:
    get:
      description: Streams player reward notifications with Server-Sent Events.
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE event stream
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ProblemDetails'
      security:
      - AuthCookieAuth: []
      summary: Stream player events
      tags:
      - me
...
schemes:
- http
- https
securityDefinitions:
  AuthCookieAuth:
    description: 'User auth cookie. Send as `Cookie: camp2026_auth=<base64url-auth-token>`.'
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"